WIP: Begin making libtww work with NTSC-U address
Added a `system::tww_ntsc_u` to directly hold the main function
definitions for the North American version as well as added features to
Cargo.toml to allow for switching between the two versions with ease.
Also, added a `.cargo/config` to specify a default target
(`powerpc-unknown-linux-gnu`) so that rustc functions properly with the
correct version of imported crates.

TODO:
	- Go through the mods and add more addresses for the NTSC-U version
	- Re-implement the custom crates and custom std
	- Possibly implement xargo for smaller build size